componente vacantes
